refactor(cli): move _iter_assistant_text_blocks to _util
